pyproject.toml
[project]
name = "garbage-classification-kubernetes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "keras-image-helper>=0.0.2",
    "numpy>=2.4.1",
    "onnxruntime>=1.23.2",
    "pillow>=12.1.0",
    "uvicorn>=0.40.0",
    "requests",
]
